standalone examples: catch up to the uint64_t API and fix what the merge audit found
Post-merge audit of master with the standalone (contiguous-array)
applications, prompted by Kale:
- simple_graph: PASSES as-is — all 15 graphs match expected.results on
  merged master, so the array-registration path survived the FoF-era
  library changes. The committed obtained.out/obtained.results were
  stray debug artifacts (run_tests regenerates and deletes them on
  success, and their numbers disagree with the gold, inviting exactly
  the confusion they caused); removed.
- prob_mesh: was broken against the uint64_t vertex-ID API (its
  getLocationFromID callback kept the old pair<int,int>(long)
  signature); updated, builds, runs to completion.
- random_graph: extern module name case fixed (unionFindLib) so it
  compiles; NOTE it is an unfinished skeleton — Main is empty and the
  program never exits. A test harness it is not.

+++ examples/prob_mesh/mesh.C
@@ -115,7 +115,8 @@ class MeshPiece : public CBase_MeshPiece {
 
     // function needed by library for quick lookup of
     // vertices location
-    static std::pair<int,int> getLocationFromID(long int vid);
+    // uint64_t vertex IDs since the FoF-campaign API change (types.h)
+    static std::pair<int,uint64_t> getLocationFromID(uint64_t vid);
         
     void initializeLibVertices() {    
         libPtr = libProxy[thisIndex].ckLocal();
@@ -170,10 +171,10 @@ class MeshPiece : public CBase_MeshPiece {
     }
 };
 
-std::pair<int,int>
-MeshPiece::getLocationFromID(long int vid) {
-    int global_y = vid % MESH_SIZE;
-    int global_x = (vid - global_y)/MESH_SIZE;
+std::pair<int,uint64_t>
+MeshPiece::getLocationFromID(uint64_t vid) {
+    int global_y = (int)(vid % MESH_SIZE);
+    int global_x = (int)((vid - global_y)/MESH_SIZE);
 
     int local_x = global_x % MESHPIECE_SIZE;
     int local_y = global_y % MESHPIECE_SIZE;
